Changed the table components to table.js
